// packages/wheelroom/src/types/wheelroom-components.ts
import { WheelroomFields } from './wheelroom-fields'

export interface WheelroomComponents {
  [componentName: string]: WheelroomComponent
}

export type AsQuery = 'page' | 'global' | 'subPage'

export interface Settings {
  /** When set boilerplate code can be generated by a plugin */
  asBoilerplate?: boolean
  /** When set asFragment files can be generated by a plugin */
  asFragment?: boolean
  /** When set a plugin can list the component a page section */
  asPageSection?: boolean
  /** When set query files can be generated by a plugin */
  asQuery?: AsQuery
  /** Can be used by a plugin to limit a query */
  queryLimit?: number
}

export interface WheelroomComponent {
  /** Fields for this component */
  fields: WheelroomFields
  /** Defines which queries are generated */
  settings: Settings
  /** Used for comparing Contentful model version with repository model version  */
  modelVersion: string
}